+++ db/engine/audit.sql
@@ -70,10 +70,10 @@ CREATE OR REPLACE FUNCTION authz.audit_check_access(
 ) RETURNS boolean
 LANGUAGE plpgsql AS $$
 DECLARE
-    v_store_id    smallint := authz._s(p_store);
-    v_user_type   smallint := authz._t(v_store_id, p_user_type);
-    v_relation    smallint := authz._r(v_store_id, p_relation);
-    v_object_type smallint := authz._t(v_store_id, p_object_type);
+    v_store_id    integer := authz._s(p_store);
+    v_user_type   integer := authz._t(v_store_id, p_user_type);
+    v_relation    integer := authz._r(v_store_id, p_relation);
+    v_object_type integer := authz._t(v_store_id, p_object_type);
     v_result      boolean;
 BEGIN
     PERFORM authz._check_namespace_access(v_store_id, v_object_type, 'can_read');
@@ -113,9 +113,9 @@ CREATE OR REPLACE FUNCTION authz.audit_list_actions(
 ) RETURNS TABLE (action text)
 LANGUAGE plpgsql AS $$
 DECLARE
-    v_store_id    smallint := authz._s(p_store);
-    v_user_type   smallint := authz._t(v_store_id, p_user_type);
-    v_object_type smallint := authz._t(v_store_id, p_object_type);
+    v_store_id    integer := authz._s(p_store);
+    v_user_type   integer := authz._t(v_store_id, p_user_type);
+    v_object_type integer := authz._t(v_store_id, p_object_type);
 BEGIN
     PERFORM authz._check_namespace_access(v_store_id, v_object_type, 'can_read');
 
@@ -173,8 +173,8 @@ CREATE OR REPLACE FUNCTION authz.audit_list_user(
 )
 LANGUAGE plpgsql STABLE AS $$
 DECLARE
-    v_store_id    smallint := authz._s(p_store);
-    v_user_type   smallint := authz._t(v_store_id, p_user_type);
+    v_store_id    integer := authz._s(p_store);
+    v_user_type   integer := authz._t(v_store_id, p_user_type);
 BEGIN
     RETURN QUERY
         SELECT a.id,
@@ -230,8 +230,8 @@ CREATE OR REPLACE FUNCTION authz.audit_list_object(
 )
 LANGUAGE plpgsql STABLE AS $$
 DECLARE
-    v_store_id    smallint := authz._s(p_store);
-    v_object_type smallint := authz._t(v_store_id, p_object_type);
+    v_store_id    integer := authz._s(p_store);
+    v_object_type integer := authz._t(v_store_id, p_object_type);
 BEGIN
     RETURN QUERY
         SELECT a.action,
